Moved viewModel logic to DataViewViewModel
